pyproject.toml
[tool.poetry]
name = "geography"
version = "0.1.0"
description = ""
authors = ["Isai Ramirez <isai.ramirez.stamaria15@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
coverage = "^6.3.2"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
django-extensions = "^3.1.5"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
